+++ packages/core/src/lib/group-info.js
@@ -1,6 +1,4 @@
-const {
-  array: { convertArrayToObject },
-} = require("@graphql-markdown/utils").scalars;
+const { convertArrayToObject } = require("@graphql-markdown/utils").array;
 
 const DEFAULT_GROUP = "Miscellaneous";
 const OPTION_REGEX =
