+++ packages/printer-legacy/src/index.js
@@ -1,31 +1,30 @@
 const {
-  isEnumType,
-  isUnionType,
-  isObjectType,
-  isScalarType,
-  isOperation,
-  getDefaultValue,
-  getTypeName,
-  getFields,
-  isDirectiveType,
-  isParametrizedField,
-  isInterfaceType,
-  getNamedType,
-  isInputType,
-  isListType,
-  isNonNullType,
-  isLeafType,
-  getRelationOfReturn,
-  getRelationOfField,
-  getRelationOfImplementation,
-  hasDirective,
-} = require("@graphql-markdown/utils").graphql;
-
-const {
   string: { toSlug, escapeMDX },
   object: { hasProperty, hasMethod },
   url: { pathUrl },
-} = require("@graphql-markdown/utils").scalars;
+  graphql: {
+    isEnumType,
+    isUnionType,
+    isObjectType,
+    isScalarType,
+    isOperation,
+    getDefaultValue,
+    getTypeName,
+    getFields,
+    isDirectiveType,
+    isParametrizedField,
+    isInterfaceType,
+    getNamedType,
+    isInputType,
+    isListType,
+    isNonNullType,
+    isLeafType,
+    getRelationOfReturn,
+    getRelationOfField,
+    getRelationOfImplementation,
+    hasDirective,
+  },
+} = require("@graphql-markdown/utils");
 
 const {
   ROOT_TYPE_LOCALE,
