+++ packages/utils/src/string.js
